+++ web/static/js/components/Game.js
