+++ web/static/js/components/Lobby.js
@@ -1,0 +1,12 @@
+import React, { Component, PropTypes } from 'react';
+import ChatContainer from '../containers/ChatContainer';
+
+export default class Lobby extends Component {
+  render() {
+    return (
+      <div>
+        <ChatContainer/>
+      </div>
+    );
+  }
+}
