+++ web/static/js/containers/ChatContainer.js
@@ -1,0 +1,13 @@
+import React, { Component, PropTypes } from 'react';
+
+import Chat from '../components/Chat';
+
+export default class ChatContainer extends Component {
+  render() {
+    return (
+      <div>
+        <Chat/>
+      </div>
+    );
+  }
+}
